requirement.txt
boto3==1.35.2
google-cloud-storage==2.18.2
pytest==8.3.2
